requirements.txt
fastapi
uvicorn
mysql-connector-python
redis
python-dotenv